First cut at transfering song:
Renamed Song to PlaylistEntry, and refactored Playlist to use this class, which
will contain information about whether the file was loaded and whether the song
was played.

Introduced PlaylistDataManager, which runs in a thread to monitor and load
playlist entires that live on remote systems.

Implemented TransferSongMessage and RequestSongMessage, for requesting and
transfering song data.

Fixed small bugs in the playlist when sending messages to update guests on play
status.

Refactored code that used Playlist to instead use accessor methods in
PlaylistService