Modified connection workflow so the host transmits a play status message to connecting guests.  Guests then set the currentEntry member of PlaylistService and dispatch a play or paused intent to notify the rest of the system.
Refactored PlaybarFragment to listen only to the intents from playlist service
(not the messages), and pull current song data from the playlist.

Refactored RemoteAudioPlayer to listen to intents from the playlist service
(instead of messages) and update the internal state accordingly.

Fixed a bug where the focus changed listener was being left registered when we
recreated an AudioPlayerWithEvents object, leading to weird play/pause behavior.