Add ACRA crash reporting to SoundStream
Adds the ACRA crash reporting to SoundStream.  Configures ACRA
to pop up a dialog in the event of a crash, asking if it's okay
to gather crash data, and get additional comments, before uploading
to the CloudAnt crash collection server.

The username and password listed are a reporting user for
the database, so it's fine to have it in public.  If worse comes
to worse, we can delete that user and generate a new one.

Also updates the dependencies file to reflect the needed ACRA
jar.