Pulls out ACRA Reporter creds into Prefs File
Pulls the credentials for the ACRA reporting
username and password into a prefs file, which
will be located in assets/app.conf.

To avoid needing to pull in an additional library,
conf file parsing is done manually, and only
supports simple syntax.  Made to gracefully fail if
Command app.conf is missing, or cannot be parsed.

Added a static reference to the AssetManager in
the CustomApp file, so that it can be referenced
elsewhere, in this case, by the prefs handling
file.